add FA support , clear code
