+++ examples/config/gpt_oss/sft_argument_gptoss_20b.json
@@ -31,6 +31,7 @@
     "zero_padding": false,
     "unified_checkpoint": true,
     "use_flash_attention": false,
+    "convert_from_hf": true,
     "lora": true,
     "lora_rank": 8,
     "rslora": false,
