+++ examples/run_finetune.py
@@ -35,6 +35,8 @@ from paddleformers.transformers import (
     Ernie4_5_MoeForCausalLMPipe,
     Ernie4_5ForCausalLM,
     Ernie4_5ForCausalLMPipe,
+    GptOssForCausalLM,
+    GptOssForCausalLMPipe,
     Llama3Tokenizer,
     LlamaForCausalLM,
     LlamaForCausalLMPipe,
@@ -65,6 +67,8 @@ flash_mask_support_list = [
     Ernie4_5ForCausalLMPipe,
     Ernie4_5_MoeForCausalLM,
     Ernie4_5_MoeForCausalLMPipe,
+    GptOssForCausalLM,
+    GptOssForCausalLMPipe,
     LlamaForCausalLM,
     LlamaForCausalLMPipe,
     Qwen2ForCausalLM,
