+++ paddleformers/transformers/gpt_oss/modeling.py
@@ -23,7 +23,6 @@ from paddle.distributed.fleet.utils.sequence_parallel_utils import ScatterOp
 from paddle.nn import functional as F
 
 from ...nn.attention.interface import ALL_ATTENTION_FUNCTIONS
-from ...nn.attention.utils import repeat_kv
 from ...nn.criterion.interface import CriterionLayer
 from ...nn.embedding import Embedding as GeneralEmbedding
 from ...nn.linear import Linear as GeneralLinear
@@ -31,7 +30,6 @@ from ...nn.lm_head import LMHead as GeneralLMHead
 from ...nn.norm import Norm as GeneralNorm
 from ...nn.pp_model import GeneralModelForCausalLMPipe
 from ...utils.log import logger
-from ...utils.tools import get_env_device
 from ..llama.modeling import get_use_casual_mask
 from ..model_outputs import MoECausalLMOutputWithPast, MoEModelOutputWithPast
 from ..model_utils import PretrainedModel, register_base_model
@@ -354,49 +352,6 @@ def apply_rotary_pos_emb(q, k, cos, sin, position_ids=None, unsqueeze_dim=1):
     return q_embed, k_embed
 
 
-def eager_attention_forward(
-    module: nn.Layer,
-    query: paddle.Tensor,
-    key: paddle.Tensor,
-    value: paddle.Tensor,
-    attention_mask: Optional[paddle.Tensor] = None,
-    dropout: float = 0.0,
-    scaling: Optional[float] = None,
-    **kwargs,
-):
-    if hasattr(module, "num_key_value_groups"):
-        num_key_value_groups = module.num_key_value_groups
-
-        key = repeat_kv(key, num_key_value_groups)
-        value = repeat_kv(value, num_key_value_groups)
-
-    perm = [0, 2, 1, 3]  # b l h d -> b h l d
-    query = paddle.transpose(x=query, perm=perm)
-    key = paddle.transpose(x=key, perm=perm)
-    value = paddle.transpose(x=value, perm=perm)
-
-    attn_weights = paddle.matmul(query, key.transpose([0, 1, 3, 2])) * scaling
-
-    if attention_mask is not None:
-
-        causal_mask = attention_mask[:, :, :, : key.shape[-2]]
-        attn_weights = attn_weights + causal_mask
-
-    sinks = module.sinks.reshape([1, -1, 1, 1]).expand([query.shape[0], -1, query.shape[-2], -1])
-
-    combined_logits = paddle.concat([attn_weights, sinks], axis=-1)
-
-    probs = F.softmax(combined_logits, axis=-1, dtype=combined_logits.dtype)
-    scores = probs[..., :-1]  # we drop the sink here
-
-    attn_weights = nn.functional.dropout(scores, p=dropout, training=module.training)
-    attn_output = paddle.matmul(attn_weights, value)  # b h l l @ b h l d -> b h l d
-    attn_output = attn_output.transpose([0, 2, 1, 3])  # b h l d -> b l h d
-    attn_output = paddle.reshape(x=attn_output, shape=[0, 0, attn_output.shape[2] * attn_output.shape[3]])
-
-    return attn_output, attn_weights
-
-
 class GptOssAttention(nn.Layer):
     """
     Multi-headed attention from 'Attention Is All You Need' paper. Modified to use sliding window attention: Longformer
@@ -484,7 +439,7 @@ class GptOssAttention(nn.Layer):
         hidden_states,
         past_key_value: Optional[Tuple[paddle.Tensor]] = None,
         attention_mask: Optional[paddle.Tensor] = None,
-        attn_mask_start_row_indices: Optional[paddle.Tensor] = None,
+        attn_mask_startend_row_indices: Optional[paddle.Tensor] = None,
         position_ids: Optional[Tuple[paddle.Tensor]] = None,
         output_attentions: bool = False,
         use_cache: bool = False,
@@ -497,7 +452,7 @@ class GptOssAttention(nn.Layer):
             hidden_states (paddle.Tensor): Input tensor [bsz, seq_len, hidden_size]
             past_key_value (Optional[Tuple[paddle.Tensor, paddle.Tensor]]): Cached key/value states
             attention_mask (Optional[paddle.Tensor]): Attention mask tensor
-            attn_mask_start_row_indices (Optional[paddle.Tensor]): Variable length attention indices
+            attn_mask_startend_row_indices (Optional[paddle.Tensor]): Variable length attention indices
             position_ids (Optional[paddle.Tensor]): Position indices for RoPE
             output_attentions (bool): Return attention weights if True
             use_cache (bool): Cache key/value states if True
@@ -536,7 +491,7 @@ class GptOssAttention(nn.Layer):
             key=key_states,
             value=value_states,
             attention_mask=attention_mask,
-            attn_mask_start_row_indices=attn_mask_start_row_indices,
+            attn_mask_startend_row_indices=attn_mask_startend_row_indices,
             sink=self.sinks,
             dropout=self.config.get("attention_dropout", 0.0) if self.training else 0.0,
             scaling=self.scaling,
@@ -592,7 +547,7 @@ class GptOssDecoderLayer(nn.Layer):
         past_key_value: Optional[Tuple[paddle.Tensor]] = None,
         use_cache: Optional[bool] = False,
         position_embedding: Optional[Tuple[paddle.Tensor, paddle.Tensor]] = None,
-        attn_mask_start_row_indices: Optional[paddle.Tensor] = None,
+        attn_mask_startend_row_indices: Optional[paddle.Tensor] = None,
         **kwargs,
     ) -> Tuple[paddle.Tensor, Optional[Tuple[paddle.Tensor, paddle.Tensor]]]:
         """
@@ -617,7 +572,7 @@ class GptOssDecoderLayer(nn.Layer):
             hidden_states=hidden_states,
             past_key_value=past_key_value,
             attention_mask=attention_mask,
-            attn_mask_start_row_indices=attn_mask_start_row_indices,
+            attn_mask_startend_row_indices=attn_mask_startend_row_indices,
             position_ids=position_ids,
             output_attentions=output_attentions,
             use_cache=use_cache,
@@ -703,99 +658,6 @@ class GptOssPreTrainedModel(PretrainedModel):
         return mappings
 
 
-def _make_sliding_window_mask(input_shape, past_key_values_length=0, window_size=5):
-    """
-    Generate a sliding window mask that restricts each position to only attend to historical positions within the window.
-    Format: [bsz, 1, tgt_seq_len, src_seq_len], where True indicates allowed attention and False indicates masking.
-    """
-    batch_size, seq_length = input_shape
-    # Total sequence length = historical sequence length + current sequence length (for generating complete mask)
-    total_length = past_key_values_length + seq_length
-
-    # Initialize mask with all False values
-    mask = paddle.zeros((seq_length, total_length), dtype=paddle.bool)
-
-    for i in range(seq_length):
-        # Absolute position of current location in the total sequence (including historical sequence)
-        current_pos = past_key_values_length + i
-        # Window start position: max(0, current position - window size + 1)
-        start = max(0, current_pos - window_size + 1)
-        # Window end position: current position (causal mask restriction, cannot exceed self)
-        end = current_pos + 1  # 切片是左闭右开，所以+1
-        # Mark window range as True (allow attention)
-        mask[i, start:end] = True
-
-    # Expand dimensions to [bsz, 1, tgt_seq_len, src_seq_len]
-    mask = mask.unsqueeze(0).unsqueeze(0)
-    # Copy to each sample in batch_size
-    mask = paddle.tile(mask, repeat_times=[batch_size, 1, 1, 1])
-    return mask
-
-
-def _prepare_decoder_attention_mask(
-    attention_mask, input_shape, past_key_values_length, dtype, sliding_window_size=None  # 新增：滑动窗口大小，None表示不启用
-):
-    # Step 1: Process input mask to generate basic expanded mask
-    if attention_mask is not None:
-        # [bsz, seq_len] -> [bsz, 1, tgt_seq_len, src_seq_len]
-        if len(attention_mask.shape) == 2:
-            expanded_attn_mask = _expand_2d_mask(attention_mask, dtype, tgt_length=input_shape[-1])
-            # When not generating in single step, need to combine causal mask and sliding window mask
-            if input_shape[-1] > 1:
-                # Generate basic causal mask (prevent future information leakage)
-                causal_mask = _make_causal_mask(input_shape, past_key_values_length=past_key_values_length)
-                # Generate sliding window mask (limit historical attention range)
-                if sliding_window_size is not None and sliding_window_size > 0:
-                    window_mask = _make_sliding_window_mask(
-                        input_shape, past_key_values_length=past_key_values_length, window_size=sliding_window_size
-                    )
-                    # Take intersection of sliding window mask and causal mask (satisfy both restrictions)
-                    combined_attention_mask = causal_mask & window_mask
-                else:
-                    combined_attention_mask = causal_mask  # Use causal mask directly when sliding window is disabled
-
-                # Combine with user-provided mask (e.g., padding mask)
-                if get_env_device() in ["npu", "mlu", "intel_hpu"]:
-                    expanded_attn_mask = expanded_attn_mask.astype("bool") & combined_attention_mask.astype("bool")
-                else:
-                    expanded_attn_mask = expanded_attn_mask & combined_attention_mask
-        # [bsz, seq_len, seq_len] -> [bsz, 1, seq_len, seq_len]
-        elif len(attention_mask.shape) == 3:
-            expanded_attn_mask = attention_mask.unsqueeze(1).astype("bool")
-        # 4D mask is used directly
-        else:
-            expanded_attn_mask = attention_mask
-    else:
-        # When no input mask, generate causal mask + sliding window mask (if enabled)
-        causal_mask = _make_causal_mask(input_shape, past_key_values_length=past_key_values_length)
-        if sliding_window_size is not None and sliding_window_size > 0:
-            window_mask = _make_sliding_window_mask(
-                input_shape, past_key_values_length=past_key_values_length, window_size=sliding_window_size
-            )
-            expanded_attn_mask = causal_mask & window_mask
-        else:
-            expanded_attn_mask = causal_mask  # Use causal mask directly when sliding window is disabled
-
-    # Step 2: Convert boolean mask to numerical mask (adapt to different devices)
-    if get_env_device() in ["npu", "mlu", "intel_hpu"]:
-        x = paddle.to_tensor(0.0, dtype="float32")
-        y = paddle.to_tensor(paddle.finfo(dtype).min, dtype="float32")
-        expanded_attn_mask = paddle.where(expanded_attn_mask.cast("bool"), x, y).astype(dtype)
-    elif get_env_device() == "xpu":
-        x = paddle.to_tensor(0.0, dtype="float32")
-        y = paddle.to_tensor(-1.7005809656952787e38, dtype="float32")
-        expanded_attn_mask = paddle.where(expanded_attn_mask.cast("bool"), x, y)
-    elif get_env_device() == "gcu":
-        min_val = paddle.finfo(dtype).min
-        x = paddle.to_tensor(0.0, dtype=dtype)
-        y = paddle.to_tensor(min_val, dtype=dtype)
-        expanded_attn_mask = paddle.where(expanded_attn_mask.cast("bool"), x, y).astype(dtype)
-    else:
-        expanded_attn_mask = paddle.where(expanded_attn_mask.cast("bool"), 0.0, paddle.finfo(dtype).min)
-        expanded_attn_mask = expanded_attn_mask.astype(dtype)
-    return expanded_attn_mask
-
-
 @register_base_model
 class GptOssModel(GptOssPreTrainedModel):
     """
@@ -851,7 +713,7 @@ class GptOssModel(GptOssPreTrainedModel):
         past_key_value: Tensor,
         use_cache: bool,
         position_embedding: Optional[Tuple[paddle.Tensor, paddle.Tensor]] = None,
-        attn_mask_start_row_indices=None,
+        attn_mask_startend_row_indices=None,
     ):
         def create_custom_forward(module):
             def custom_forward(*inputs):
@@ -869,7 +731,7 @@ class GptOssModel(GptOssPreTrainedModel):
             past_key_value,
             use_cache,
             position_embedding,
-            attn_mask_start_row_indices,
+            attn_mask_startend_row_indices,
             # use_reentrant=self.config.recompute_use_reentrant,
         )
 
@@ -887,7 +749,7 @@ class GptOssModel(GptOssPreTrainedModel):
         output_hidden_states: Optional[bool] = None,
         output_router_logits: Optional[bool] = None,
         return_dict: Optional[bool] = None,
-        attn_mask_start_row_indices=None,
+        attn_mask_startend_row_indices=None,
         **kwargs,
     ) -> Union[Tuple, MoEModelOutputWithPast]:
         output_attentions = output_attentions if output_attentions is not None else self.config.output_attentions
@@ -932,7 +794,7 @@ class GptOssModel(GptOssPreTrainedModel):
             inputs_embeds = ScatterOp.apply(inputs_embeds)
 
         # embed positions
-        if attn_mask_start_row_indices is not None or get_use_casual_mask():
+        if attn_mask_startend_row_indices is not None or get_use_casual_mask():
             attention_mask = None
         else:
             # [bs, seq_len]
@@ -944,7 +806,7 @@ class GptOssModel(GptOssPreTrainedModel):
             causal_mask_mapping = {}
 
             # full_attention
-            causal_mask = _prepare_decoder_attention_mask(
+            causal_mask = self._prepare_decoder_attention_mask(
                 attention_mask=attention_mask,
                 input_shape=(batch_size, seq_length),
                 past_key_values_length=cache_length,
@@ -955,7 +817,7 @@ class GptOssModel(GptOssPreTrainedModel):
             causal_mask_mapping["full_attention"] = causal_mask
 
             # sliding_attention
-            causal_mask = _prepare_decoder_attention_mask(
+            causal_mask = self._prepare_decoder_attention_mask(
                 attention_mask=attention_mask,
                 input_shape=(batch_size, seq_length),
                 past_key_values_length=cache_length,
@@ -989,7 +851,7 @@ class GptOssModel(GptOssPreTrainedModel):
                     layer_module=decoder_layer,
                     hidden_states=hidden_states,
                     attention_mask=causal_mask_mapping[decoder_layer.attention_type],
-                    attn_mask_start_row_indices=attn_mask_start_row_indices,
+                    attn_mask_startend_row_indices=attn_mask_startend_row_indices,
                     position_ids=position_ids,
                     output_attentions=output_attentions,
                     output_router_logits=output_router_logits,
@@ -1001,7 +863,7 @@ class GptOssModel(GptOssPreTrainedModel):
                 layer_outputs = decoder_layer(
                     hidden_states=hidden_states,
                     attention_mask=causal_mask_mapping[decoder_layer.attention_type],
-                    attn_mask_start_row_indices=attn_mask_start_row_indices,
+                    attn_mask_startend_row_indices=attn_mask_startend_row_indices,
                     position_ids=position_ids,
                     output_attentions=output_attentions,
                     output_router_logits=output_router_logits,
